Add 2018 Open ELO calculation
Clea up the 2018 data and pass it through the ELO calculation.

Updated K to (players/(players-1)) + 1, which comes out to
about 2. This seems to keep the ELOs from spiraling out of
controll. This allows the K to be used for 1200 players in
the Open data and the 40 in the Games.

Signed-off-by: Marc Jones <marcj303@gmail.com>